api/pyproject.toml
[tool.poetry]
name = "es-assignment"
version = "0.1.0"
description = ""
authors = ["Tomás Mendes <you@example.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.111.0"
uvicorn = "^0.29.0"
pydantic = {extras = ["dotenv", "email"], version = "^2.7.1"}
httpx = "^0.27.0"
pydantic-settings = "^2.2.1"
email-validator = "^2.1.1"
boto3 = "^1.34.110"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}

[tool.poetry.group.dev.dependencies]
black = "^24.4.2"
flake8 = "^7.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
